\name{benchmark}
\alias{benchmark}
\title{
  AML31 benchmarking tool
}
\description{
  Takes a 5-column single-nucleotide variant list generated by running
  your favorite tool on the AML31 data. Produces some informative plots
  and statistics, along with a table listing which variants were valid
  in both the "Platinum" and "Gold" truth sets.
}
\usage{
    benchmark(variantFile, tablePrefix, plotFile)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{variantFile}{
    Single nucleotide variants in 5-column format: 1) chromosome 2)
    start position 3) stop position, 4) reference allele 5) variant allele
  }
  \item{plotFile}{
    filename to which plots will be output (in PDF output)
  }
  \item{tablePrefix}{
    filename prefix to which labelled variants will be output (in tsv
    format). Two files will be created: <prefix>.gold and
    <prefix>.platinum
  }
}
